+++ libakumuli/CMakeLists.txt
@@ -4,21 +4,20 @@ add_library(akumuli SHARED
     ../include/akumuli.h
     ../include/akumuli_def.h
     log_iface.h
-    volume.h
+    storage_engine/volume.h
     storage.h
     page.h
     util.h
     sort.h
     sequencer.h
     cursor.h
-    compression.h
-    compression.cpp
+    storage_engine/compression.h
+    storage_engine/compression.cpp
     internal_cursor.h
-    compression.h
     metadatastorage.h
     stringpool.h
     log_iface.cpp
-    volume.cpp
+    storage_engine/volume.cpp
     storage.cpp
     seriesparser.cpp
     page.cpp
@@ -46,7 +45,7 @@ add_library(akumuli SHARED
     query_processing/limiter.cpp
 )
 
-include_directories(./libakumuli)
+include_directories(.)
 
 include(CppcheckTargets)
 add_cppcheck(akumuli UNUSED_FUNCTIONS STYLE POSSIBLE_ERROR FORCE)
