+++ libakumuli/storage_engine/blockstore.h
