+++ libakumuli/storage_engine/compression.cpp
